Update ID for Eall entity and add logic to front-end for liked item

+++ back-end/src/main/java/com/devminds/rentify/entity/Role.java
@@ -19,13 +19,10 @@ import lombok.Data;
 @Table(name = "user_role")
 
 public class Role {
-
-
-
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Id
     @Column(name = "id")
-    private int id;
+    private long id;
 
     @Enumerated(EnumType.STRING)
     @Column(name = "role_name")
